growthzi_backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
cohere==5.14.2
colorama==0.4.6
distro==1.9.0
Django==3.1.12
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.7
exceptiongroup==1.2.2
fastavro==1.10.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
jiter==0.9.0
mongoengine==0.29.1
multidict==6.3.2
openai==0.28.0
packaging==24.2
propcache==0.3.1
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.9.0
pymongo==3.11.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.2.4
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.19.0
